+++ packages/terra-clinical-item-display/src/terra-dev-site/test/clinical-item-display/display/IconItemDisplay.test.jsx
@@ -3,4 +3,18 @@ import IconSuccess from 'terra-icon/lib/icon/IconSuccess';
 
 import ItemDisplay from '../../../../ItemDisplay';
 
-export default () => <ItemDisplay text="Test Text" icon={<IconSuccess />} id="ItemDisplay" />;
+const testStyle = {
+  padding: '10px 0',
+};
+
+const testText = 'Wrapping Test Text Wrapping Test Text Wrapping Test Text Wrapping Test Text Wrapping Test Text Wrapping Test Text Wrapping Test Text Wrapping Test Text Wrapping Test Text Wrapping Test Text';
+
+export default () => (
+  <div style={testStyle}>
+    <ItemDisplay text={`Center / default - ${testText}`} icon={<IconSuccess />} id="test-icon-center" />
+    <br />
+    <ItemDisplay text={`Top - ${testText}`} icon={<IconSuccess />} id="test-icon-top" iconAlignment="top" />
+    <br />
+    <ItemDisplay text={`Inline - ${testText}`} icon={<IconSuccess />} id="test-icon-inline" iconAlignment="inline" />
+  </div>
+);