+++ packages/terra-clinical-item-display/CHANGELOG.md
@@ -3,6 +3,12 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* iconAlignment prop
+
+### Fixed 
+* Added missing primary text style and corresponding --terra-clinical-item-display-primary-color variable
+* Added enumeration for valid text style values
 
 3.4.0 - (June 11, 2019)
 ----------