+++ packages/terra-clinical-onset-picker/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Changed
 * Updated jest snapshots.
 
+### Added
+* Added enumeration for valid age unit, granularity, precision, and precision set values
+
 4.3.0 - (June 11, 2019)
 ----------
 ### Removed